devs_sources/python/MQTT_Client/requirements.txt
paho-mqtt
python-periphery
